requirements.txt
beautifulsoup4==4.4.1
lxml==3.4.4
Pillow==3.0.0
python-dateutil==2.6.0
requests==2.8.1
six==1.10.0
